LayoutTests/fast/text-autosizing/ios/idempotentmode/idempotent-autosizing-after-changing-initial-scale-expected.txt

At initial scale 1
none	auto
8px	8px
12px	12px
14px	14px
16px	16px
20px	20px

After changing initial scale to 0.5
none	auto
8px	13px
12px	17px
14px	19px
16px	19px
20px	20px

After changing initial scale to 0.75
none	auto
8px	10px
12px	14px
14px	16px
16px	16px
20px	20px

After changing initial scale back to 1
none	auto
8px	8px
12px	12px
14px	14px
16px	16px
20px	20px

After changing initial scale to 1.25
none	auto
8px	8px
12px	12px
14px	14px
16px	16px
20px	20px
